requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.9.3
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
aiosqlite==0.20.0
    # via -r requirements.in
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via langchain-core
appdirs==1.4.4
    # via yfinance
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via yfinance
certifi==2024.2.2
    # via
    #   curl-cffi
    #   requests
cffi==1.16.0
    # via curl-cffi
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   duckduckgo-search
    #   nltk
curl-cffi==0.6.2
    # via duckduckgo-search
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
duckduckgo-search==5.1.0
    # via -r requirements.in
frozendict==2.4.0
    # via yfinance
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
html5lib==1.1
    # via yfinance
idna==3.6
    # via
    #   anyio
    #   requests
    #   yarl
joblib==1.3.2
    # via nltk
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
jsonschema==4.21.1
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
langchain==0.1.9
    # via -r requirements.in
langchain-community==0.0.29
    # via langchain
langchain-core==0.1.33
    # via
    #   langchain
    #   langchain-community
langsmith==0.1.31
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lxml==5.1.0
    # via
    #   duckduckgo-search
    #   yfinance
mako==1.3.2
    # via alembic
markupsafe==2.1.5
    # via mako
marshmallow==3.21.1
    # via dataclasses-json
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multitasking==0.0.11
    # via yfinance
mypy-extensions==1.0.0
    # via typing-inspect
nltk==3.8.1
    # via -r requirements.in
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
    #   pandas
    #   yfinance
orjson==3.9.15
    # via langsmith
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
pandas==2.2.1
    # via yfinance
peewee==3.17.1
    # via yfinance
pycparser==2.21
    # via cffi
pydantic==2.6.2
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.16.3
    # via pydantic
pytelegrambotapi==4.16.1
    # via telebot
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via
    #   pandas
    #   yfinance
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-core
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2023.12.25
    # via nltk
requests==2.31.0
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
    #   pytelegrambotapi
    #   telebot
    #   yfinance
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   html5lib
    #   python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.28
    # via
    #   -r requirements.in
    #   alembic
    #   langchain
    #   langchain-community
telebot==0.0.5
    # via -r requirements.in
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tqdm==4.66.2
    # via nltk
typing-extensions==4.10.0
    # via
    #   aiosqlite
    #   alembic
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
webencodings==0.5.1
    # via html5lib
yarl==1.9.4
    # via aiohttp
yfinance==0.2.36
    # via -r requirements.in
